// frontend/src/models/user.ts
export interface User {

    // User Information
    _id: number;
    user_first_name: string;
    user_last_name: string;
    user_prefix: string;
    username: string;
    account_type: string;
    bio: string;
    contact_phone: string;
    contact_email: string;
    password: number;
    password_conf: number;

}